GH-236: 'Node' and 'masq' implementation, finalized

+++ node/src/accountant/payable_dao.rs
@@ -67,7 +67,7 @@ pub trait PayableDao: Debug + Send {
 
     fn top_records(&self, minimum_amount: u64, maximum_age: u64) -> Vec<PayableAccount>;
 
-    fn total(&self) -> u64;
+    fn total(&self) -> i64;
 }
 
 pub trait PayableDaoFactory {
@@ -228,7 +228,7 @@ impl PayableDao for PayableDaoReal {
         .collect()
     }
 
-    fn total(&self) -> u64 {
+    fn total(&self) -> i64 {
         let mut stmt = self
             .conn
             .prepare("select sum(balance) from payable")
@@ -236,7 +236,7 @@ impl PayableDao for PayableDaoReal {
         match stmt.query_row([], |row| {
             let total_balance_result: Result<i64, rusqlite::Error> = row.get(0);
             match total_balance_result {
-                Ok(total_balance) => Ok(total_balance as u64),
+                Ok(total_balance) => Ok(total_balance),
                 Err(e)
                     if e == rusqlite::Error::InvalidColumnType(
                         0,
@@ -244,7 +244,7 @@ impl PayableDao for PayableDaoReal {
                         Type::Null,
                     ) =>
                 {
-                    Ok(0u64)
+                    Ok(0)
                 }
                 Err(e) => panic!(
                     "Database is corrupt: PAYABLE table columns and/or types: {:?}",
